+++ testing/util/block_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/primitives"
 	"github.com/prysmaticlabs/prysm/v5/encoding/bytesutil"
 	ethpbv1 "github.com/prysmaticlabs/prysm/v5/proto/eth/v1"
-	ethpbv2 "github.com/prysmaticlabs/prysm/v5/proto/eth/v2"
 	ethpbalpha "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/testing/require"
 )
@@ -216,28 +215,6 @@ func TestHydrateV1SignedBeaconBlock_NoError(t *testing.T) {
 	require.NoError(t, err)
 }
 
-func TestHydrateV2AltairSignedBeaconBlock_NoError(t *testing.T) {
-	b := &ethpbv2.SignedBeaconBlockAltair{}
-	b = HydrateV2AltairSignedBeaconBlock(b)
-	_, err := b.HashTreeRoot()
-	require.NoError(t, err)
-	_, err = b.Message.HashTreeRoot()
-	require.NoError(t, err)
-	_, err = b.Message.Body.HashTreeRoot()
-	require.NoError(t, err)
-}
-
-func TestHydrateV2BellatrixSignedBeaconBlock_NoError(t *testing.T) {
-	b := &ethpbv2.SignedBeaconBlockBellatrix{}
-	b = HydrateV2BellatrixSignedBeaconBlock(b)
-	_, err := b.HashTreeRoot()
-	require.NoError(t, err)
-	_, err = b.Message.HashTreeRoot()
-	require.NoError(t, err)
-	_, err = b.Message.Body.HashTreeRoot()
-	require.NoError(t, err)
-}
-
 func TestHydrateSignedBeaconBlockAltair_NoError(t *testing.T) {
 	b := &ethpbalpha.SignedBeaconBlockAltair{}
 	b = HydrateSignedBeaconBlockAltair(b)
@@ -274,33 +251,6 @@ func TestHydrateBlindedBeaconBlockBellatrix_NoError(t *testing.T) {
 func TestHydrateBlindedBeaconBlockBodyBellatrix_NoError(t *testing.T) {
 	b := &ethpbalpha.BlindedBeaconBlockBodyBellatrix{}
 	b = HydrateBlindedBeaconBlockBodyBellatrix(b)
-	_, err := b.HashTreeRoot()
-	require.NoError(t, err)
-}
-
-func TestHydrateV2SignedBlindedBeaconBlockBellatrix_NoError(t *testing.T) {
-	b := &ethpbv2.SignedBlindedBeaconBlockBellatrix{}
-	b = HydrateV2SignedBlindedBeaconBlockBellatrix(b)
-	_, err := b.HashTreeRoot()
-	require.NoError(t, err)
-	_, err = b.Message.HashTreeRoot()
-	require.NoError(t, err)
-	_, err = b.Message.Body.HashTreeRoot()
-	require.NoError(t, err)
-}
-
-func TestHydrateV2BlindedBeaconBlockBellatrix_NoError(t *testing.T) {
-	b := &ethpbv2.BlindedBeaconBlockBellatrix{}
-	b = HydrateV2BlindedBeaconBlockBellatrix(b)
-	_, err := b.HashTreeRoot()
-	require.NoError(t, err)
-	_, err = b.Body.HashTreeRoot()
-	require.NoError(t, err)
-}
-
-func TestHydrateV2BlindedBeaconBlockBodyBellatrix_NoError(t *testing.T) {
-	b := &ethpbv2.BlindedBeaconBlockBodyBellatrix{}
-	b = HydrateV2BlindedBeaconBlockBodyBellatrix(b)
 	_, err := b.HashTreeRoot()
 	require.NoError(t, err)
 }