+++ testing/util/merge.go
@@ -1,7 +1,6 @@
 package util
 
 import (
-	v2 "github.com/prysmaticlabs/prysm/v5/proto/eth/v2"
 	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 )
 
@@ -13,11 +12,6 @@ func NewBeaconBlockBellatrix() *ethpb.SignedBeaconBlockBellatrix {
 // NewBlindedBeaconBlockBellatrix creates a blinded beacon block with minimum marshalable fields.
 func NewBlindedBeaconBlockBellatrix() *ethpb.SignedBlindedBeaconBlockBellatrix {
 	return HydrateSignedBlindedBeaconBlockBellatrix(&ethpb.SignedBlindedBeaconBlockBellatrix{})
-}
-
-// NewBlindedBeaconBlockBellatrixV2 creates a blinded beacon block with minimum marshalable fields.
-func NewBlindedBeaconBlockBellatrixV2() *v2.SignedBlindedBeaconBlockBellatrix {
-	return HydrateV2SignedBlindedBeaconBlockBellatrix(&v2.SignedBlindedBeaconBlockBellatrix{})
 }
 
 // NewBeaconBlockCapella creates a beacon block with minimum marshalable fields.
@@ -58,9 +52,4 @@ func NewBeaconBlockContentsElectra() *ethpb.SignedBeaconBlockContentsElectra {
 // NewBlindedBeaconBlockElectra creates a blinded beacon block with minimum marshalable fields.
 func NewBlindedBeaconBlockElectra() *ethpb.SignedBlindedBeaconBlockElectra {
 	return HydrateSignedBlindedBeaconBlockElectra(&ethpb.SignedBlindedBeaconBlockElectra{})
-}
-
-// NewBlindedBeaconBlockCapellaV2 creates a blinded beacon block with minimum marshalable fields.
-func NewBlindedBeaconBlockCapellaV2() *v2.SignedBlindedBeaconBlockCapella {
-	return HydrateV2SignedBlindedBeaconBlockCapella(&v2.SignedBlindedBeaconBlockCapella{})
 }